+++ packages/rasa/src/models/rasa-msg-log.aggregate.ts
