+++ packages/server/package.json
@@ -26,6 +26,7 @@
     "@senses-chat/operator-common": "0.1.0",
     "@senses-chat/operator-database": "0.1.0",
     "@senses-chat/operator-events": "0.1.0",
+    "@senses-chat/operator-rasa": "0.1.0",
     "@senses-chat/wx-sdk": "0.1.0",
     "body-parser-xml": "^2.0.3",
     "class-transformer": "^0.5.1",
