+++ packages/server/src/app.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { RouteModule } from '@senses-chat/operator-events';
+import { RasaModule } from '@senses-chat/operator-rasa';
 
 import { ConfigModule } from './modules';
 import { WechatModule } from './wechat';
 import { WxkfModule } from './wxkf';
-import { RasaModule } from './rasa';
 import { ApiModule } from './api';
 
 import { AppController } from './app.controller';
