+++ packages/server/src/modules/config.module.ts
@@ -1,7 +1,6 @@
 import { DynamicModule } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 
-import rasaConfig from 'src/config/rasa';
 import serverConfig from 'src/config/server';
 import wxkfConfig from 'src/config/wxkf';
 import wx3pConfig from 'src/config/wx3p';
@@ -12,7 +11,6 @@ export const Module: DynamicModule = ConfigModule.forRoot({
   // mimic behaviors from nextjs
   envFilePath: [`.env.${ENV}.local`, `.env.${ENV}`, `.env.local`, '.env'],
   load: [
-    rasaConfig,
     serverConfig,
     wxkfConfig,
     wx3pConfig,
